DONE: Edit + Delete function + Display Messages
Added distinct PHP code files to implement the user delete from database and user details edit. Deletion will refresh the customer list via forced redirect to admin page. 
Additionally, the edit-user file gets the id through the href="" method and retrieves it through the $_GET["id"] variable, as is stated in the documentation as well.
Finally, the display message function hints the admin to use the message board, due to difficulties bypassing the Bootstrap pill tab list(the selected container would reset on page refresh) I chose to stick with in the end .